Move derivation of TLS 1.2 cipher pair into ConnectionSecrets method
